# backend/openapi/openapi.selfhosted.yaml
openapi: 3.0.3
info:
  title: SlugBase API (self-hosted)
  description: |
    HTTP API for a self-hosted SlugBase instance. Replace the server URL with your deployment base URL.

    **Authentication:** Most routes accept either a session cookie (browser) after login, or `Authorization: Bearer <api_token>` from `/api/tokens`. Personal API tokens do **not** require interactive MFA when a user has TOTP enabled (same model as GitHub PATs: possession of the secret is the factor).

    **MFA (TOTP):** Optional second factor after password or OIDC. Step-up uses a short-lived `slugbase.mfa_pending` cookie until `POST /api/auth/mfa/verify` succeeds. Management routes require a normal session plus CSRF where applicable.

    **CSRF:** For cookie-based sessions, non-GET requests generally require `X-CSRF-Token` from `GET /api/csrf-token` first. Some auth and password-reset routes are exempt (see code: `app-factory.ts`).
  version: 1.0.0
servers:
  - url: /
    description: Instance root (e.g. https://bookmarks.example.com)
...
paths:
...
  /api/admin/users/{id}:
    get:
      tags: [AdminUsers]
      summary: Get user
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: User
    put:
      tags: [AdminUsers]
      summary: Update user
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Updated
    delete:
      tags: [AdminUsers]
      summary: Delete user
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Deleted
  /api/admin/users/{id}/teams:
    get:
      tags: [AdminUsers]
      summary: Teams for user
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Teams
...
  /api/admin/teams/{id}:
    get:
      tags: [AdminTeams]
      summary: Get team
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Team
    put:
      tags: [AdminTeams]
      summary: Update team
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Updated
    delete:
      tags: [AdminTeams]
      summary: Delete team
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Deleted
  /api/admin/teams/{id}/members:
    post:
      tags: [AdminTeams]
      summary: Add team member
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: OK
  /api/admin/teams/{id}/members/{userId}:
    delete:
      tags: [AdminTeams]
      summary: Remove team member
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
...
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: slugbase.sid
      description: Session cookie after login
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API token
  parameters:
    IdPath:
      in: path
      name: id
      required: true
      schema:
        type: string
...
security:
  - cookieAuth: []
  - bearerAuth: []